requirements.txt
langchain
langchain_community
langchain_text_splitters
langchain_core
python_dotenv
streamlit
langchain_groq
langchain_huggingface
faiss-cpu
pypdf
wikipedia
arxiv
duckduckgo-search
youtube_transcript_api
pytube
unstructured
validators==0.28.1
numexpr